// engineGateway/src/generated/engineGateway.ts
export interface addOrUpdateGatewayRequest {
  secret: string;
  order?: Order | undefined;
}

export interface deleteOrderGatewayRequest {
  secret: string;
  orderId: number;
}

export interface gatewayResponse {
  status: string;
  message: string;
}

function createBaseaddOrUpdateGatewayRequest(): addOrUpdateGatewayRequest {
  return { secret: "", order: undefined };
}

export const addOrUpdateGatewayRequest: MessageFns<addOrUpdateGatewayRequest> = {
  encode(message: addOrUpdateGatewayRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.secret !== "") {
      writer.uint32(10).string(message.secret);
    }
    if (message.order !== undefined) {
      Order.encode(message.order, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): addOrUpdateGatewayRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseaddOrUpdateGatewayRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.secret = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.order = Order.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): addOrUpdateGatewayRequest {
    return {
      secret: isSet(object.secret) ? globalThis.String(object.secret) : "",
      order: isSet(object.order) ? Order.fromJSON(object.order) : undefined,
    };
  },

  toJSON(message: addOrUpdateGatewayRequest): unknown {
    const obj: any = {};
    if (message.secret !== "") {
      obj.secret = message.secret;
    }
    if (message.order !== undefined) {
      obj.order = Order.toJSON(message.order);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<addOrUpdateGatewayRequest>, I>>(base?: I): addOrUpdateGatewayRequest {
    return addOrUpdateGatewayRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<addOrUpdateGatewayRequest>, I>>(object: I): addOrUpdateGatewayRequest {
    const message = createBaseaddOrUpdateGatewayRequest();
    message.secret = object.secret ?? "";
    message.order = (object.order !== undefined && object.order !== null) ? Order.fromPartial(object.order) : undefined;
    return message;
  },
};
// ...
export const gatewayResponse: MessageFns<gatewayResponse> = {
  encode(message: gatewayResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== "") {
      writer.uint32(10).string(message.status);
    }
    if (message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): gatewayResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasegatewayResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): gatewayResponse {
    return {
      status: isSet(object.status) ? globalThis.String(object.status) : "",
      message: isSet(object.message) ? globalThis.String(object.message) : "",
    };
  },

  toJSON(message: gatewayResponse): unknown {
    const obj: any = {};
    if (message.status !== "") {
      obj.status = message.status;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<gatewayResponse>, I>>(base?: I): gatewayResponse {
    return gatewayResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<gatewayResponse>, I>>(object: I): gatewayResponse {
    const message = createBasegatewayResponse();
    message.status = object.status ?? "";
    message.message = object.message ?? "";
    return message;
  },
};
// ...
export interface EngineGatewayServiceServer extends UntypedServiceImplementation {
  /**
   * rpc UpdateOrder (addOrUpdateGatewayRequest) returns (gatewayResponse);
   * rpc DeleteOrder (deleteOrderGatewayRequest) returns (gatewayResponse);
   */
  addOrder: handleUnaryCall<addOrUpdateGatewayRequest, gatewayResponse>;
}
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}