// engineGateway/src/generated/engineRequest.ts
export enum ActionType {
	ADD_ORDER = 0,
	CANCEL_ORDER = 1,
	UPDATE_ORDER = 2,
	UNRECOGNIZED = -1,
}
// ...
export interface engineRequest {
	/** used mainly for CANCEL_ORDER */
	orderId: number;
	/** for quick routing to correct orderbook */
	ticker: string;
	/** ADD, CANCEL, UPDATE */
	action: ActionType;
	/** from order.proto, null / empty for CANCEL */
	order: Order | undefined;
}

function createBaseengineRequest(): engineRequest {
	return { orderId: 0, ticker: "", action: 0, order: undefined };
}

export const engineRequest: MessageFns<engineRequest> = {
	encode(message: engineRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
		if (message.orderId !== 0) {
			writer.uint32(8).int32(message.orderId);
		}
		if (message.ticker !== "") {
			writer.uint32(18).string(message.ticker);
		}
		if (message.action !== 0) {
			writer.uint32(24).int32(message.action);
		}
		if (message.order !== undefined) {
			Order.encode(message.order, writer.uint32(34).fork()).join();
		}
		return writer;
	},

	decode(input: BinaryReader | Uint8Array, length?: number): engineRequest {
		const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
		const end = length === undefined ? reader.len : reader.pos + length;
		const message = createBaseengineRequest();
		while (reader.pos < end) {
			const tag = reader.uint32();
			switch (tag >>> 3) {
				case 1: {
					if (tag !== 8) {
						break;
					}

					message.orderId = reader.int32();
					continue;
				}
				case 2: {
					if (tag !== 18) {
						break;
					}

					message.ticker = reader.string();
					continue;
				}
				case 3: {
					if (tag !== 24) {
						break;
					}

					message.action = reader.int32() as any;
					continue;
				}
				case 4: {
					if (tag !== 34) {
						break;
					}

					message.order = Order.decode(reader, reader.uint32());
					continue;
				}
			}
			if ((tag & 7) === 4 || tag === 0) {
				break;
			}
			reader.skip(tag & 7);
		}
		return message;
	},

	fromJSON(object: any): engineRequest {
		return {
			orderId: isSet(object.orderId) ? globalThis.Number(object.orderId) : 0,
			ticker: isSet(object.ticker) ? globalThis.String(object.ticker) : "",
			action: isSet(object.action) ? actionTypeFromJSON(object.action) : 0,
			order: isSet(object.order) ? Order.fromJSON(object.order) : undefined,
		};
	},

	toJSON(message: engineRequest): unknown {
		const obj: any = {};
		if (message.orderId !== 0) {
			obj.orderId = Math.round(message.orderId);
		}
		if (message.ticker !== "") {
			obj.ticker = message.ticker;
		}
		if (message.action !== 0) {
			obj.action = actionTypeToJSON(message.action);
		}
		if (message.order !== undefined) {
			obj.order = Order.toJSON(message.order);
		}
		return obj;
	},

	create<I extends Exact<DeepPartial<engineRequest>, I>>(base?: I): engineRequest {
		return engineRequest.fromPartial(base ?? ({} as any));
	},
	fromPartial<I extends Exact<DeepPartial<engineRequest>, I>>(object: I): engineRequest {
		const message = createBaseengineRequest();
		message.orderId = object.orderId ?? 0;
		message.ticker = object.ticker ?? "";
		message.action = object.action ?? 0;
		message.order = (object.order !== undefined && object.order !== null) ? Order.fromPartial(object.order) : undefined;
		return message;
	},
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
	: T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
	: T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
	: T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
	: Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
	: P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
	return value !== null && value !== undefined;
}

export interface MessageFns<T> {
	encode(message: T, writer?: BinaryWriter): BinaryWriter;
	decode(input: BinaryReader | Uint8Array, length?: number): T;
	fromJSON(object: any): T;
	toJSON(message: T): unknown;
	create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
	fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}