// engineGateway/src/generated/orderEvent.ts
export enum OrderEventType {
	ORDER_ADDED = 0,
	ORDER_CANCELLED = 1,
	ORDER_UPDATED = 2,
	TRADE_EXECUTABLE = 3,
	ORDER_ADD_FAILED = 4,
	ORDER_UPDATE_FAILED = 5,
	ORDER_CANCEL_FAILED = 6,
	TRADE_EXECUTE_FAILED = 7,
	TRADE_EXECUTED = 8,
	UNRECOGNIZED = -1,
}
// ...
export interface OrderEvent {
	event: OrderEventType;
	/** from order.proto */
	order: Order | undefined;
	message: string;
}

function createBaseOrderEvent(): OrderEvent {
	return { event: 0, order: undefined, message: "" };
}

export const OrderEvent: MessageFns<OrderEvent> = {
	encode(message: OrderEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
		if (message.event !== 0) {
			writer.uint32(8).int32(message.event);
		}
		if (message.order !== undefined) {
			Order.encode(message.order, writer.uint32(18).fork()).join();
		}
		if (message.message !== "") {
			writer.uint32(26).string(message.message);
		}
		return writer;
	},

	decode(input: BinaryReader | Uint8Array, length?: number): OrderEvent {
		const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
		const end = length === undefined ? reader.len : reader.pos + length;
		const message = createBaseOrderEvent();
		while (reader.pos < end) {
			const tag = reader.uint32();
			switch (tag >>> 3) {
				case 1: {
					if (tag !== 8) {
						break;
					}

					message.event = reader.int32() as any;
					continue;
				}
				case 2: {
					if (tag !== 18) {
						break;
					}

					message.order = Order.decode(reader, reader.uint32());
					continue;
				}
				case 3: {
					if (tag !== 26) {
						break;
					}

					message.message = reader.string();
					continue;
				}
			}
			if ((tag & 7) === 4 || tag === 0) {
				break;
			}
			reader.skip(tag & 7);
		}
		return message;
	},

	fromJSON(object: any): OrderEvent {
		return {
			event: isSet(object.event) ? orderEventTypeFromJSON(object.event) : 0,
			order: isSet(object.order) ? Order.fromJSON(object.order) : undefined,
			message: isSet(object.message) ? globalThis.String(object.message) : "",
		};
	},

	toJSON(message: OrderEvent): unknown {
		const obj: any = {};
		if (message.event !== 0) {
			obj.event = orderEventTypeToJSON(message.event);
		}
		if (message.order !== undefined) {
			obj.order = Order.toJSON(message.order);
		}
		if (message.message !== "") {
			obj.message = message.message;
		}
		return obj;
	},

	create<I extends Exact<DeepPartial<OrderEvent>, I>>(base?: I): OrderEvent {
		return OrderEvent.fromPartial(base ?? ({} as any));
	},
	fromPartial<I extends Exact<DeepPartial<OrderEvent>, I>>(object: I): OrderEvent {
		const message = createBaseOrderEvent();
		message.event = object.event ?? 0;
		message.order = (object.order !== undefined && object.order !== null) ? Order.fromPartial(object.order) : undefined;
		message.message = object.message ?? "";
		return message;
	},
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
	: T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
	: T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
	: T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
	: Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
	: P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
	return value !== null && value !== undefined;
}

export interface MessageFns<T> {
	encode(message: T, writer?: BinaryWriter): BinaryWriter;
	decode(input: BinaryReader | Uint8Array, length?: number): T;
	fromJSON(object: any): T;
	toJSON(message: T): unknown;
	create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
	fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}